+++ lib/src/drag_over_position.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/widgets.dart';
 
-class DockingDrag extends ChangeNotifier {
+class DragOverPosition extends ChangeNotifier {
   bool _enable = false;
 
   bool get enable => _enable;