+++ lib/src/internal/widgets/docking_item_widget.dart
@@ -1,5 +1,5 @@
 import 'package:docking/src/docking_buttons_builder.dart';
-import 'package:docking/src/docking_drag.dart';
+import 'package:docking/src/drag_over_position.dart';
 import 'package:docking/src/internal/widgets/draggable_config_mixin.dart';
 import 'package:docking/src/internal/widgets/drop/content_wrapper.dart';
 import 'package:docking/src/internal/widgets/drop/drop_feedback_widget.dart';
@@ -19,13 +19,14 @@ class DockingItemWidget extends StatefulWidget {
   DockingItemWidget(
       {Key? key,
       required this.layout,
-      required this.dockingDrag,
+      required this.dragOverPosition,
       required this.item,
       this.onItemSelection,
       this.onItemClose,
       this.itemCloseInterceptor,
       this.dockingButtonsBuilder,
-      required this.maximizable})
+      required this.maximizable,
+      required this.draggable})
       : super(key: key);
 
   final DockingLayout layout;
@@ -35,7 +36,8 @@ class DockingItemWidget extends StatefulWidget {
   final ItemCloseInterceptor? itemCloseInterceptor;
   final DockingButtonsBuilder? dockingButtonsBuilder;
   final bool maximizable;
-  final DockingDrag dockingDrag;
+  final DragOverPosition dragOverPosition;
+  final bool draggable;
 
   @override
   State<StatefulWidget> createState() => DockingItemWidgetState();
@@ -84,7 +86,8 @@ class DockingItemWidgetState extends State<DockingItemWidget>
           content: content,
           closable: widget.item.closable,
           leading: widget.item.leading,
-          buttons: buttons)
+          buttons: buttons,
+          draggable: widget.draggable)
     ];
     TabbedViewController controller = TabbedViewController(tabs);
 
@@ -103,18 +106,19 @@ class DockingItemWidgetState extends State<DockingItemWidget>
         tabCloseInterceptor: _tabCloseInterceptor,
         onTabClose: _onTabClose,
         controller: controller,
-        onDraggableBuild:
-            (TabbedViewController controller, int tabIndex, TabData tabData) {
-          return buildDraggableConfig(
-              dockingDrag: widget.dockingDrag, tabData: tabData);
-        },
+        onDraggableBuild: widget.draggable
+            ? (TabbedViewController controller, int tabIndex, TabData tabData) {
+                return buildDraggableConfig(
+                    dockingDrag: widget.dragOverPosition, tabData: tabData);
+              }
+            : null,
         contentBuilder: (context, tabIndex) => ItemContentWrapper(
             listener: _updateActiveDropPosition,
             layout: widget.layout,
             dockingItem: widget.item,
             child: controller.tabs[tabIndex].content!),
-        onBeforeDropAccept: _onBeforeDropAccept);
-    if (widget.dockingDrag.enable) {
+        onBeforeDropAccept: widget.draggable ? _onBeforeDropAccept : null);
+    if (widget.draggable && widget.dragOverPosition.enable) {
       return DropFeedbackWidget(
           dropPosition: _activeDropPosition, child: tabbedView);
     }