+++ lib/src/internal/widgets/docking_tabs_widget.dart
@@ -1,7 +1,7 @@
 import 'dart:math' as math;
 
 import 'package:docking/src/docking_buttons_builder.dart';
-import 'package:docking/src/docking_drag.dart';
+import 'package:docking/src/drag_over_position.dart';
 import 'package:docking/src/internal/widgets/draggable_config_mixin.dart';
 import 'package:docking/src/internal/widgets/drop/content_wrapper.dart';
 import 'package:docking/src/internal/widgets/drop/drop_feedback_widget.dart';
@@ -19,14 +19,15 @@ class DockingTabsWidget extends StatefulWidget {
   DockingTabsWidget(
       {Key? key,
       required this.layout,
-      required this.dockingDrag,
+      required this.dragOverPosition,
       required this.dockingTabs,
       this.onItemSelection,
       this.onItemClose,
       this.itemCloseInterceptor,
       this.dockingButtonsBuilder,
       required this.maximizableTab,
-      required this.maximizableTabsArea})
+      required this.maximizableTabsArea,
+      required this.draggable})
       : super(key: key);
 
   final DockingLayout layout;
@@ -37,7 +38,8 @@ class DockingTabsWidget extends StatefulWidget {
   final DockingButtonsBuilder? dockingButtonsBuilder;
   final bool maximizableTab;
   final bool maximizableTabsArea;
-  final DockingDrag dockingDrag;
+  final DragOverPosition dragOverPosition;
+  final bool draggable;
 
   @override
   State<StatefulWidget> createState() => DockingTabsWidgetState();
@@ -86,7 +88,8 @@ class DockingTabsWidgetState extends State<DockingTabsWidget>
           closable: child.closable,
           keepAlive: child.globalKey != null,
           leading: child.leading,
-          buttons: buttons));
+          buttons: buttons,
+          draggable: widget.draggable));
     });
     TabbedViewController controller = TabbedViewController(tabs);
     controller.selectedIndex =
@@ -104,19 +107,20 @@ class DockingTabsWidgetState extends State<DockingTabsWidget>
           }
         },
         tabCloseInterceptor: _tabCloseInterceptor,
-        onDraggableBuild:
-            (TabbedViewController controller, int tabIndex, TabData tabData) {
-          return buildDraggableConfig(
-              dockingDrag: widget.dockingDrag, tabData: tabData);
-        },
+        onDraggableBuild: widget.draggable
+            ? (TabbedViewController controller, int tabIndex, TabData tabData) {
+                return buildDraggableConfig(
+                    dockingDrag: widget.dragOverPosition, tabData: tabData);
+              }
+            : null,
         onTabClose: _onTabClose,
         contentBuilder: (context, tabIndex) => TabsContentWrapper(
             listener: _updateActiveDropPosition,
             layout: widget.layout,
             dockingTabs: widget.dockingTabs,
             child: controller.tabs[tabIndex].content!),
-        onBeforeDropAccept: _onBeforeDropAccept);
-    if (widget.dockingDrag.enable) {
+        onBeforeDropAccept: widget.draggable ? _onBeforeDropAccept : null);
+    if (widget.draggable && widget.dragOverPosition.enable) {
       return DropFeedbackWidget(
           dropPosition: _activeDropPosition, child: tabbedView);
     }